# WIP Let most error handling functions and shared object init hooks be flagged as cold and be optimized for size to improve CPU frontend throughput and branch prediction