--- a/array.c
+++ b/array.c
@@ -5943,7 +5943,7 @@ rb_ary_sum(int argc, VALUE *argv, VALUE ary)
     v = finish_exact_sum(n, r, v, argc!=0);
     return v;
 
-  not_exact:
+  not_exact: COLDLABEL
     v = finish_exact_sum(n, r, v, i!=0);
 
     if (RB_FLOAT_TYPE_P(e)) {
@@ -6250,6 +6250,7 @@ rb_ary_sum(int argc, VALUE *argv, VALUE ary)
  *
  */
 
+COLDFUNC(void Init_Array(void));
 void
 Init_Array(void)
 {