--- a/ext/continuation/continuation.c
+++ b/ext/continuation/continuation.c
@@ -1,8 +1,9 @@
 
 #include "ruby/ruby.h"
 
-void ruby_Init_Continuation_body(void);
+COLDFUNC(void ruby_Init_Continuation_body(void));
 
+COLDFUNC(void Init_continuation(void));
 void
 Init_continuation(void)
 {