--- a/ext/fiddle/closure.h
+++ b/ext/fiddle/closure.h
@@ -3,6 +3,6 @@
 
 #include <fiddle.h>
 
-void Init_fiddle_closure(void);
+COLDFUNC(void Init_fiddle_closure(void));
 
 #endif