--- a/ext/gdbm/gdbm.c
+++ b/ext/gdbm/gdbm.c
@@ -84,8 +84,8 @@ static VALUE rb_cGDBM, rb_eGDBMError, rb_eGDBMFatalError;
 #define MY_BLOCK_SIZE (2048)
 #define MY_FATAL_FUNC rb_gdbm_fatal
 
-NORETURN(static void rb_gdbm_fatal(const char *msg));
-NORETURN(static void closed_dbm(void));
+NORETURN(COLDFUNC(static void rb_gdbm_fatal(const char *msg)));
+NORETURN(COLDFUNC(static void closed_dbm(void)));
 
 static void
 rb_gdbm_fatal(const char *msg)
@@ -1224,6 +1224,7 @@ fgdbm_reject(VALUE obj)
     return rb_hash_delete_if(fgdbm_to_hash(obj));
 }
 
+COLDFUNC(void Init_gdbm(void));
 void
 Init_gdbm(void)
 {