--- a/ext/io/wait/wait.c
+++ b/ext/io/wait/wait.c
@@ -42,7 +42,7 @@
 static VALUE io_ready_p _((VALUE io));
 static VALUE io_wait_readable _((int argc, VALUE *argv, VALUE io));
 static VALUE io_wait_writable _((int argc, VALUE *argv, VALUE io));
-void Init_wait _((void));
+COLDFUNC(void Init_wait _((void)));
 
 static struct timeval *
 get_timeout(int argc, VALUE *argv, struct timeval *timerec)
@@ -243,6 +243,7 @@ io_wait_readwrite(int argc, VALUE *argv, VALUE io)
  * IO wait methods
  */
 
+COLDFUNC(void Init_wait(void));
 void
 Init_wait(void)
 {