--- a/ext/openssl/ossl_asn1.h
+++ b/ext/openssl/ossl_asn1.h
@@ -57,6 +57,6 @@ extern VALUE cASN1Sequence, cASN1Set;                /* CONSTRUCTIVE      */
 
 ASN1_TYPE *ossl_asn1_get_asn1type(VALUE);
 
-void Init_ossl_asn1(void);
+COLDFUNC(void Init_ossl_asn1(void));
 
 #endif