--- a/ext/openssl/ossl_cipher.h
+++ b/ext/openssl/ossl_cipher.h
@@ -15,6 +15,6 @@ extern VALUE eCipherError;
 
 const EVP_CIPHER *ossl_evp_get_cipherbyname(VALUE);
 VALUE ossl_cipher_new(const EVP_CIPHER *);
-void Init_ossl_cipher(void);
+COLDFUNC(void Init_ossl_cipher(void));
 
 #endif /* _OSSL_CIPHER_H_ */