--- a/template/limits.c.tmpl
+++ b/template/limits.c.tmpl
@@ -51,6 +51,7 @@
 # include <float.h>
 #endif
 
+COLDFUNC(void Init_limits(void));
 void
 Init_limits(void)
 {