I distilled the feedback given on https://github.com/ruby/ruby/pull/1915 for a while and also concluded that:

* Sprinkling branch predictor hints for callsites of specific functions leads to bloat, complexity and also a standard / contract very difficult to enforce
* It's essentially a higher level version of what modern compilers already do.

Also for [ruby 3x3](https://rubykaigi.org/2017/presentations/codefolio.html) I feel we should look towards "hardware empathy" more - help the compiler generate better code with hints, take a pass through the memory layout of internal structures ( [SOA vs OAS](https://en.wikipedia.org/wiki/AOS_and_SOA) ), data locality and possibility also have a look @ how modern game engines push hardware to get the maximum work done per game loop (essentially a VM tick in Ruby land).

This lead me to the `hot` and `cold` [attributes](https://gcc.gnu.org/onlinedocs/gcc/Common-Function-Attributes.html#Common-Function-Attributes) supported by `GCC` and `clang`, extensively used in the [kernel](https://github.com/torvalds/linux/blob/ca04b3cca11acbaf904f707f2d9ca9654d7cc226/include/linux/compiler-gcc.h#L191-L206), PHP (especially the swift 7.x releases) and other projects. 

### Implementation details

This PR only focuses on the `cold` [cold](https://gcc.gnu.org/onlinedocs/gcc-4.6.4/gcc/Function-Attributes.html), which works in the following way (from GCC docs):

```
The cold attribute is used to inform the compiler that a function is unlikely executed. The function is optimized for size rather than speed and on many targets it is placed into special subsection of the text section so all cold functions appears close together improving code locality of non-cold parts of program. The paths leading to call of cold functions within code are marked as unlikely by the branch prediction mechanism. It is thus useful to mark functions used to handle unlikely conditions, such as perror, as cold to improve optimization of hot functions that do call marked functions in rare occasions.
When profile feedback is available, via -fprofile-use, hot functions are automatically detected and this attribute is ignored.
```
By declaring a function as `cold` when defined we get the following benefits:

* No excessive `unlikely` macros sprinkled about
* We get the branch prediction hints for free @ every callsite
* No-op on platforms that does not support the attributes
* Size optimization of cold functions with a smaller footprint in the instruction cache
* Therefore CPU frontend throughput increases due to a lower ratio of instruction cache misses and a lower ITLB overhead.
* This effect can further be amplified in future work with the `hot` attribute

This pattern also translates well to `Init_` hooks of extensions and other shared ruby libs.  Therefore this PR also defines `Init_` entry point functions for core extensions as cold - smaller icache as they are generally referenced from within requires which should not be in a hot path and thus fine to optimize for size. And only ever called once.

GCC 5+ also supports `cold` [labels](https://gcc.gnu.org/onlinedocs/gcc/Label-Attributes.html), to inform the compiler that a particular label is unlikely to be executed. There's several error and similar labels in the codebase and I was targeting those primarily.

Binary sizes this branch:

```
lourens@CarbonX1:~/src/ruby/ruby$ size ruby
   text	   data	    bss	    dec	    hex	filename
3462225	  21088	  71344	3554657	 363d61	ruby
```

VS trunk (`2832` bytes smaller text segment)

```
lourens@CarbonX1:~/src/ruby/trunk$ size ruby
   text	   data	    bss	    dec	    hex	filename
3465057	  21088	  71344	3557489	 364871	ruby
```

The cold functions are placed in the special section `.text.unlikely` section below (number 6)

```
lourens@CarbonX1:~/src/ruby/ruby$ readelf -S error.o
There are 32 section headers, starting at offset 0x85848:

Section Headers:
  [Nr] Name              Type             Address           Offset
       Size              EntSize          Flags  Link  Info  Align
  [ 0]                   NULL             0000000000000000  00000000
       0000000000000000  0000000000000000           0     0     0
  [ 1] .text             PROGBITS         0000000000000000  00000040
       000000000000666d  0000000000000000  AX       0     0     16
  [ 2] .rela.text        RELA             0000000000000000  000397d0
       000000000000acf8  0000000000000018   I      29     1     8
  [ 3] .data             PROGBITS         0000000000000000  000066ad
       0000000000000000  0000000000000000  WA       0     0     1
  [ 4] .bss              NOBITS           0000000000000000  000066c0
       00000000000010d8  0000000000000000  WA       0     0     32
  [ 5] .rodata.str1.1    PROGBITS         0000000000000000  000066c0
       000000000000097b  0000000000000001 AMS       0     0     1
  [ 6] .text.unlikely    PROGBITS         0000000000000000  0000703b
       0000000000001f0b  0000000000000000  AX       0     0     1
  [ 7] .rela.text.unlike RELA             0000000000000000  000444c8
       0000000000002190  0000000000000018   I      29     6     8
  [ 8] .rodata.str1.8    PROGBITS         0000000000000000  00008f48
       000000000000023f  0000000000000001 AMS       0     0     8
  [ 9] .rodata           PROGBITS         0000000000000000  000091a0
       0000000000000226  0000000000000000   A       0     0     32
 --- truncated for brevity ---
```

Proof of the error specific functions migrated to the `.text.unlikely` section.

```
lourens@CarbonX1:~/src/ruby/ruby$ ld -M error.o

--- only a small snippet for brevity ----

.text           0x0000000000400120     0x857d
 *(.text.unlikely .text.*_unlikely .text.unlikely.*)
 .text.unlikely
                0x0000000000400120     0x1f25 error.o
                0x0000000000400ac2                rb_syntax_error_append
                0x0000000000400c9f                rb_warning_warn
                0x0000000000400d1d                rb_compile_warn
                0x0000000000400ddb                rb_compile_warning
                0x0000000000400e9c                rb_warn
                0x0000000000400f54                rb_enc_warn
                0x000000000040100d                rb_warning
                0x00000000004010c8                rb_bug
                0x00000000004011a5                rb_bug_context
                0x0000000000401283                rb_bug_errno
                0x00000000004012e7                rb_async_bug_errno
                0x000000000040142a                rb_report_bug_valist
                0x000000000040147a                rb_exc_new
                0x00000000004014aa                rb_exc_new_cstr
                0x00000000004014c7                rb_exc_new_str
                0x000000000040152b                rb_name_error
                0x00000000004015e9                rb_name_error_str
                0x000000000040174f                rb_name_err_new
                0x000000000040177d                rb_nomethod_err_new
                0x00000000004017f6                rb_key_err_new
                0x0000000000401869                rb_raise
                0x000000000040190a                rb_exc_set_backtrace
                0x000000000040190f                rb_invalid_str
                0x0000000000401933                rb_loaderror
                0x00000000004019db                rb_loaderror_with_path
                0x0000000000401a80                rb_notimplement
                0x0000000000401aa9                rb_fatal
                0x0000000000401b4c                ruby_deprecated_internal_feature
                0x0000000000401b66                rb_syserr_new_str
                0x0000000000401b8b                rb_syserr_new
                0x0000000000401c41                rb_syserr_fail
                0x0000000000401c52                rb_syserr_fail_str
                0x0000000000401c63                rb_sys_fail
                0x0000000000401c74                rb_sys_fail_str
                0x0000000000401c85                rb_mod_sys_fail
                0x0000000000401cac                rb_mod_sys_fail_str
                0x0000000000401cd3                rb_mod_syserr_fail
                0x0000000000401cfc                rb_mod_syserr_fail_str
                0x0000000000401d25                rb_sys_warning
                0x0000000000401ded                rb_sys_enc_warning
                0x0000000000401eb8                rb_syserr_enc_warning
                0x0000000000401f70                rb_load_fail
                0x0000000000401fae                rb_error_frozen
                0x0000000000401fca                rb_error_frozen_object
                0x0000000000402044                rb_error_untrusted

lourens@CarbonX1:~/src/ruby/ruby$ ld -M vm.o

--- only a small snippet for brevity ----
.text           0x0000000000400120    0x1e1bf
 *(.text.unlikely .text.*_unlikely .text.unlikely.*)
 .text.unlikely
                0x0000000000400120      0x767 vm.o
                0x0000000000400192                rb_f_notimplement
                0x0000000000400572                rb_error_arity
                0x000000000040063c                rb_make_no_method_exception
                0x0000000000400807                rb_vm_localjump_error
                0x0000000000400818                rb_vm_make_jump_tag_but_local_jump

lourens@CarbonX1:~/src/ruby/ruby$ ld -M eval.o

--- only a small snippet for brevity ----

.text           0x0000000000400120     0x5302
 *(.text.unlikely .text.*_unlikely .text.unlikely.*)
 .text.unlikely
                0x0000000000400120     0x131f eval.o
                0x0000000000400b04                rb_error_write
                0x0000000000400c85                rb_ec_error_print
                0x0000000000400fbf                rb_ec_setup_exception
                0x000000000040100e                rb_exc_raise
                0x000000000040104e                rb_print_undef
                0x0000000000401131                rb_print_undef_str
                0x0000000000401190                rb_print_inaccessible
                0x0000000000401273                rb_exc_fatal
                0x00000000004012b3                rb_interrupt
                0x00000000004012cf                rb_make_exception
                0x00000000004013f8                rb_errinfo
                0x0000000000401404                rb_set_errinfo

```

Some `Init_` functions relocated to the cold section as well:

```
lourens@CarbonX1:~/src/ruby/ruby$ ld -M eval.o

--- only a small snippet for brevity ----

.text           0x0000000000400120    0x10769
 *(.text.unlikely .text.*_unlikely .text.unlikely.*)
 .text.unlikely
                0x0000000000400120      0xbfe array.o
                0x0000000000400139                Init_Array
 *(.text.exit .text.exit.*)
 *(.text.startup .text.startup.*)
 *(.text.hot .text.hot.*)
 *(.text .stub .text.* .gnu.linkonce.t.*)
 *fill*         0x0000000000400d1e        0x2 
 .text          0x0000000000400d20     0xfb69 array.o
                0x0000000000401950                rb_ary_rassoc
                0x0000000000401a20                rb_ary_includes
                0x0000000000401aa0                rb_ary_assoc

lourens@CarbonX1:~/src/ruby/ruby$ ld -M hash.o

--- only a small snippet for brevity ----

.text           0x0000000000400120     0xae07
 *(.text.unlikely .text.*_unlikely .text.unlikely.*)
 .text.unlikely
                0x0000000000400120      0xece hash.o
                0x0000000000400156                Init_Hash
 *(.text.exit .text.exit.*)
 *(.text.startup .text.startup.*)
 *(.text.hot .text.hot.*)
 *(.text .stub .text.* .gnu.linkonce.t.*)
 *fill*         0x0000000000400fee        0x2 
 .text          0x0000000000400ff0     0x9f37 hash.o
                0x00000000004010c0                rb_hash_size
                0x00000000004011d0                rb_hash_compare_by_id_p
                0x0000000000401340                rb_hash_has_key
                0x0000000000401700                rb_hash_keys

```

### What this PR does NOT do, and future work

* Completely ignores the `hot` attribute for now as it's a completely different concern and references future work.
* Implement support for [Intel VTune tracing technology APIs](https://software.intel.com/en-us/vtune-amplifier-help-instrumentation-and-tracing-technology-apis) for better insights on what bytecode is actually doing with the hardware. Primarily to determine where else we're frontend bound still in the VM
* Opt in hugepages support to reduce [TLB](https://en.wikipedia.org/wiki/Translation_lookaside_buffer) overhead for the code segment (map it to a few 2MB pages instead of several 4k ones). Possibly useful for object layouts / the ruby heap for large Rails applications (not database server territory, but there may be some benefit)
* Look into the benefits of [link time optimization](https://en.wikipedia.org/wiki/Interprocedural_optimization)

### Downsides

* Mostly GCC specific (possibly works nicely with clang too, no tests with it yet)
* Higher rate of mispredictions and stalls for code that rely on exceptions and errors for flow control (not a great pattern to use anyways)
* Keeping tabs on / awareness of the `COLDFUNC` and `COLDLABEL` macros - didn't seem to burden kernel, PHP and other core devs very much ...

### Benchmarks

Benchmarks, especially instructions p/s, instruction cache performance and branch prediction in progress

#### Branches

```
Samples: 17K of event 'branches', Event count (approx.): 8707306167
Overhead  Command         Shared Object      Symbol
  66,22%  ruby            ruby               [.] vm_exec_core
   5,24%  ruby            ruby               [.] vm_setivar.isra.230
   3,88%  ruby            ruby               [.] vm_call_iseq_setup_normal_0start_0params_0locals
   2,65%  ruby            ruby               [.] vm_call_cfunc
   1,75%  ruby            ruby               [.] rb_class_of
   1,58%  ruby            ruby               [.] rb_ary_push
   1,16%  ruby            ruby               [.] vm_call_iseq_setup_normal_0start_0params_2locals
   1,08%  ruby            ruby               [.] ary_ensure_room_for_push
   0,88%  ruby            ruby               [.] vm_call_iseq_setup_normal_0start_0params_1locals
   0,84%  ruby            ruby               [.] invoke_iseq_block_from_c
   0,80%  ruby            ruby               [.] vm_call_iseq_setup
   0,74%  ruby            ruby               [.] rb_ary_transpose
   0,58%  ruby            ruby               [.] rb_ary_rotate_bang
   0,55%  ruby            ruby               [.] vm_call_iseq_setup_normal_0start_1params_1locals
   0,55%  ruby            ruby               [.] rb_vm_exec
```

VS trunk

```
Samples: 16K of event 'branches', Event count (approx.): 8714006421
Overhead  Command         Shared Object      Symbol
  68,15%  ruby            ruby               [.] vm_exec_core
   4,98%  ruby            ruby               [.] vm_setivar.isra.230
   3,12%  ruby            ruby               [.] vm_call_iseq_setup_normal_0start_0params_0locals
   2,74%  ruby            ruby               [.] vm_call_cfunc
   1,76%  ruby            ruby               [.] rb_class_of
   1,73%  ruby            ruby               [.] rb_ary_push
   0,95%  ruby            ruby               [.] ary_ensure_room_for_push
   0,88%  ruby            ruby               [.] vm_call_iseq_setup_normal_0start_0params_2locals
   0,75%  ruby            ruby               [.] vm_call_iseq_setup
   0,72%  ruby            ruby               [.] rb_ary_transpose
   0,68%  ruby            ruby               [.] invoke_iseq_block_from_c
   0,61%  ruby            ruby               [.] rb_vm_exec
   0,61%  ruby            ruby               [.] vm_call_iseq_setup_normal_0start_0params_1locals
   0,57%  ruby            ruby               [.] rb_yield
   0,50%  ruby            ruby               [.] opt_eq_func
   0,50%  ruby            ruby               [.] rb_ary_rotate_bang
   0,49%  ruby            ruby               [.] vm_call_iseq_setup_normal_0start_1params_1locals
   0,47%  ruby            ruby               [.] vm_call_opt_send
```

#### Branch misses

```
Samples: 15K of event 'branch-misses', Event count (approx.): 43797451
Overhead  Command         Shared Object      Symbol
  73,94%  ruby            ruby               [.] vm_exec_core
   6,77%  ruby            ruby               [.] vm_call_cfunc
   5,31%  ruby            ruby               [.] vm_setivar.isra.230
   2,69%  ruby            ruby               [.] vm_call_iseq_setup_normal_0start_0params_2locals
   2,41%  ruby            ruby               [.] vm_call_iseq_setup_normal_0start_0params_0locals
   1,88%  ruby            ruby               [.] vm_call_iseq_setup_normal_0start_1params_1locals
   1,03%  ruby            ruby               [.] rb_class_of
   0,53%  ruby            ruby               [.] ruby_yyparse
```

VS trunk

```
Samples: 15K of event 'branch-misses', Event count (approx.): 45440490
Overhead  Command         Shared Object      Symbol
  80,53%  ruby            ruby               [.] vm_exec_core
   7,00%  ruby            ruby               [.] vm_setivar.isra.230
   2,68%  ruby            ruby               [.] vm_call_cfunc
   1,65%  ruby            ruby               [.] rb_class_of
   1,11%  ruby            ruby               [.] vm_call_iseq_setup_normal_0start_0params_0locals
   0,70%  ruby            ruby               [.] vm_call_iseq_setup_normal_0start_0params_2locals
   0,64%  ruby            ruby               [.] vm_call_iseq_setup_normal_0start_1params_1locals
   0,59%  ruby            ruby               [.] ruby_yyparse

```
### CPU frontend utilization improvements

Lower instruction TLB overhead and instruction cache misses. The goal with this PR is to improve throughput of instructions for the CPU frontend as ruby seems to stall the CPU significantly for some workloads. Trying to reduce frontend stalls with relocating cold functions and branch predictor hints of code executing less is the primary objective here. An expected side effect would be noticing backend optimization opportunities more easily as the frontend throughput increases

VTune output from `benchmark/app_erb.yml`

![frontend_cold](https://user-images.githubusercontent.com/379/44204858-4c085100-a14c-11e8-86b8-d87fcb5e4985.png)

VS trunk

![frontend_trunk](https://user-images.githubusercontent.com/379/44204870-4f9bd800-a14c-11e8-9bee-14c8ad8d3a7d.png)

#### Observations from the above

* Less stalls and a larger fraction utilized for work
* Better CPI(cycles per instruction) rate (lower is better)
* Less frontend bound (not supplying instructions fast enough to the backend). 30% is still very high.
* No icache misses
* Ditto for ITLB misses
* Higher rate of retired instructions (more is better)
* Slight improvement in branch prediction
* This branch is more backend bound (expected because the frontend stalls less)

#### Initial benchmarks

General improvement in [optcarrot](https://github.com/mame/optcarrot) FPS

```
lourens@CarbonX1:~/src/optcarrot$ benchmark-driver -e "cold::~/src/ruby/ruby/ruby -I~/src/ruby/ruby/lib -I~/src/ruby/ruby/. -I~/src/ruby/ruby/.ext/x86_64-linux" -e "trunk::~/src/ruby/trunk/ruby -I~/src/ruby/trunk/lib -I~/src/ruby/trunk/. -I~/src/ruby/trunk/.ext/x86_64-linux" --repeat-count 1 --repeat-result average -vv benchmark.yml
cold: ruby 2.6.0dev (2018-08-16 cold-error-paths 64392) [x86_64-linux]
last_commit=Let most error handling functions and shared object init hooks be flagged as cold and be optimized for size
…